Join HoeffdingTreeRegressor nodes
