+++ src/skmultiflow/trees/hoeffding_adaptive_tree_regressor.py
@@ -197,7 +197,7 @@ class HoeffdingAdaptiveTreeRegressor(HoeffdingTreeRegressor):
             self.sum_of_attribute_values += weight * X
             self.sum_of_attribute_squares += weight * X * X
         except ValueError:
-            
+
             self.sum_of_attribute_values = weight * X
             self.sum_of_attribute_squares = weight * X * X
